Use $target_file in *_get variables, not $name
When a user specifies the $target parameter in a staging::file defined
type, the intention is to fetch the file from $source and place it at
$target. When $name is used in the $http_get variants, the retrieved
file won't actually be placed at the $target location. This commit makes
it so that curl and wget will always place the retrieved file in the
$target_file location. This should not impact the behavior when a user
doesn't specify $target, but should fix the unexpected behavior when
they do.